Make CFFI bindings out-of-line and fix deps
