+++ developer_requirements.txt
@@ -1,0 +1,4 @@
+sphinx_rtd_theme
+sphinx>=1.3
+scikit-image
+scipy
