+++ docs/conf.py
@@ -33,7 +33,7 @@ sys.path.insert(0,
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.autodoc', 'sphinxcontrib.napoleon',
+extensions = ['sphinx.ext.autodoc', 'sphinx.ext.napoleon',
               'sphinx.ext.mathjax']
 
 # Add any paths that contain templates here, relative to this directory.
