+++ docs/index.rst
@@ -1,7 +1,7 @@
 Proximal total-variation operators
 **********************************
 
-.. automodule:: prox_tv._prox_tv
+.. automodule:: prox_tv
 
 Function reference
 ==================
@@ -10,29 +10,29 @@ One dimensional total variation problems
 ----------------------------------------
 
 
-.. autofunction:: prox_tv._prox_tv.tv1_1d
+.. autofunction:: prox_tv.tv1_1d
 
-.. autofunction:: prox_tv._prox_tv.tv1w_1d
+.. autofunction:: prox_tv.tv1w_1d
 
-.. autofunction:: prox_tv._prox_tv.tv2_1d
+.. autofunction:: prox_tv.tv2_1d
 
-.. autofunction:: prox_tv._prox_tv.tvp_1d
+.. autofunction:: prox_tv.tvp_1d
 
 
 Two dimensional total variation problems
 ----------------------------------------
 
 
-.. autofunction:: prox_tv._prox_tv.tv1_2d
+.. autofunction:: prox_tv.tv1_2d
 
-.. autofunction:: prox_tv._prox_tv.tv1w_2d
+.. autofunction:: prox_tv.tv1w_2d
 
-.. autofunction:: prox_tv._prox_tv.tvp_2d
+.. autofunction:: prox_tv.tvp_2d
 
 Generalized total variation problems
 ------------------------------------
 
-.. autofunction:: prox_tv._prox_tv.tvgen
+.. autofunction:: prox_tv.tvgen
 
 Examples
 ========
