+++ tox.ini
@@ -1,0 +1,5 @@
+[tox]
+envlist=py35,py27
+[testenv]
+deps=nose
+commands=nosetests
